[water] lowering read/write with mappings (#995)
Make read/write lowering handle permutations in the mapping by swapping
the order of index expressions following the inverse permutation
(because the mapping is memory->value and the index expressions are for
value). Practically this means reordering `ordered_syms` since the MLIR
dictionary attribute order is always alphabetical. Note that this is
partly inconsistent with the python counterpart as it only permutes
start expressions, which it get can away with by not using step/stride
after permutation.

Change `ordered_sym` to track value shape instead of memory shape; it
wasn't relevant until now because the two were verifiably identical, but
will matter with permutation mappings. The order of index in pywave is
that of the `indexing_dims`, which is taken from the non-memory part of
the mapping (output for reads, input for writes), which is now verified
to match the value shape when present. Update the distributed shape
resolution pass accordingly. Add verifiers to guarantee this.

Change the logic of generating transfer_read/write, which should be
dependent on the which dimension is vectorized according to the index
expression _not_ the strides of the memref, which transfer read should
be capable of handling by itself. Be defensive about it and bail out
from patterns should memref layout have non-consecutive strides. We
don't have such cases currently.

We should consider revising the read/write lowerings holistically since
pywave doesn't seem to use transfer_read/write and the logic around
mapping and masking is questionable.

---------

+++ water/lib/Dialect/Wave/IR/WaveUtils.cpp
@@ -45,11 +45,11 @@ wave::getUncollapsedVectorShape(llvm::ArrayRef<wave::WaveSymbolAttr> shape,
   });
 }
 
-std::optional<uint64_t>
+std::optional<int64_t>
 wave::getPositionOfVectorizedDim(llvm::ArrayRef<wave::WaveSymbolAttr> shape,
                                  DictionaryAttr indexDict,
                                  wave::WaveHyperparameterAttr hyper) {
-  uint64_t bestIdx = static_cast<uint64_t>(-1);
+  int64_t bestIdx = -1;
   std::optional<int64_t> bestSize; // largest constant size seen so far
   for (auto [i, size] :
        llvm::enumerate(getUncollapsedVectorShape(shape, indexDict, hyper))) {
@@ -60,7 +60,7 @@ wave::getPositionOfVectorizedDim(llvm::ArrayRef<wave::WaveSymbolAttr> shape,
       bestIdx = i;
     }
   }
-  assert(bestIdx != static_cast<uint64_t>(-1));
+  assert(bestIdx != -1);
   return bestIdx;
 }
